advanced_level/arquitectura_limpia/src/app/core/security.py
# core/security.py

from datetime import datetime, timedelta

from jose import jwt

SECRET_KEY = "supersecret"
ALGORITHM = "HS256"


def create_access_token(data: dict):
    to_encode = data.copy()
    to_encode.update({"exp": datetime.utcnow() + timedelta(minutes=30)})

    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)


def decode_token(token: str):
    return jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
